Updated User model, New Account View Controller, and fixed issues with project file and main story board
